Refactorizar y mejorar los informes de inventario y ventas
Refactorice métodos y clases para mejorar la claridad y la funcionalidad. Se agregó generación de tablas para inventario, productos más vendidos e informes de ventas de proveedores. Tipos de datos actualizados y errores lógicos corregidos, mejorando la precisión del procesamiento de datos.

+++ src/co/edu/uptc/dto/ReporteInventarioDTO.java
@@ -6,7 +6,7 @@ public class ReporteInventarioDTO {
 	private double precioBase;
 	private double totalPrecioVenta;
 	private double totalImpuesto;
-	private int totalComisiones;
+	private double totalComisiones;
 	private double totalGanancias;
 
 	public ReporteInventarioDTO() {}
@@ -43,11 +43,11 @@ public class ReporteInventarioDTO {
 		this.totalImpuesto = totalImpuesto;
 	}
 
-	public int getTotalComisiones() {
+	public double getTotalComisiones() {
 		return totalComisiones;
 	}
 
-	public void setTotalComisiones(int totalComisiones) {
+	public void setTotalComisiones(double totalComisiones) {
 		this.totalComisiones = totalComisiones;
 	}
